Use unsigned integer types for non-negative counts (nsub, nsim, nall, nSize)
Change rx_solve struct fields nsub, nsim, nall from int to uint32_t so
that they cannot hold semantically invalid negative values and the type
communicates intent to future maintainers.  Fix rxSimOmega to use
uint32_t for its nStud/nSub parameters and R_xlen_t for allocation
expressions to prevent signed integer overflow when nSub * nStud would
exceed INT_MAX.  Change all nSizeLong intermediates from int64_t to
uint64_t so that the overflow detection uses unsigned arithmetic.
Change rxSolve_t::nSize to uint32_t and update gparsCovSetup to accept
uint32_t nsub.  Update all downstream consumers (par_solve.cpp,
rxode2_df.cpp, approx.cpp, rx2api.c) with explicit casts where needed
to avoid new signed/unsigned comparison warnings.  Add an overflow
guard in rxSimThetaOmega that throws an informative error when
nSub * nStud > INT_MAX instead of segfaulting in the allocator.

Test: the new test in test-issue-999.R confirms that calling rxSolve
with nSub=46342, nStud=46342 (product > INT_MAX) now produces an
informative 'too large' error rather than the previous crash with
'negative length vectors are not allowed'.

+++ src/rxode2_df.cpp
@@ -131,7 +131,7 @@ extern "C" SEXP rxode2_df(int doDose0, int doTBS) {
   int nkeep  = rx->nKeepF;
   int nlhs = op->nlhs;
   int nobs = rx->nobs - rx->nevid9;
-  int nsim = rx->nsim;
+  int nsim = (int)rx->nsim;
   int nall = rx->nall - rx->nevid9;
   int errNcol = rxGetErrsNcol();
   int errNrow = rxGetErrsNrow();
@@ -171,7 +171,7 @@ extern "C" SEXP rxode2_df(int doDose0, int doTBS) {
     dullRate=1, dullDur=1,
     dullSS=1, dullIi=1;
   int csub = 0, evid = 0;
-  int nsub = rx->nsub;
+  int nsub = (int)rx->nsub;
   IntegerVector rmState = rxStateIgnore(op->modNamePtr);
   int nPrnState =0;
   int i, j;
@@ -240,7 +240,7 @@ extern "C" SEXP rxode2_df(int doDose0, int doTBS) {
       // (Rf_errorcall)(R_NilValue, "%s", _("'alag(.)'/'rate(.)'/'dur(.)' cannot depend on the state values"));
     }
     if (nidCols == 0){
-      for (int solveid = 0; solveid < rx->nsub * rx->nsim; solveid++){
+      for (int solveid = 0; solveid < (int)(rx->nsub * rx->nsim); solveid++){
         rx_solving_options_ind *indE = &(rx->subjects[solveid]);
         if (indE->err != 0) {
           printErr(indE->err, indE->id);
